add check to only delete tslint.json if present
- addresses PR's #1 & #3 that did more than just this fix

+++ scripts/postinstall.ts
@@ -122,6 +122,7 @@ fs.writeFileSync(GULPFILE_FILEPATH, gulpFileData.replace(/build.initialize\(requ
 /**
  *
  * STEP 4: DELETE TSLINT.JSON
+ * > only if present
  *
 */
 console.log('');
@@ -130,5 +131,7 @@ console.log(`INFO: Deleting tslint.info unnecessary file`);
 
 const TSLINT_FILEPATH = path.resolve(path.join(projectPath, 'tslint.json'));
 
-// delete file
-fs.unlinkSync(TSLINT_FILEPATH);
+// delete file (if present)
+if (fs.existsSync(TSLINT_FILEPATH)) {
+  fs.unlinkSync(TSLINT_FILEPATH);
+}
